Adds a flag to fail fast to Q2
This commit add a flag "-ff" to the Q2 command line arguments.

Q2 will shutdown immediately if a bean fails to deploy or start and the
flag is on. This change doesn't prevent Q2 form renaming the file.

Two test are added to Q2Test to verify that this change doesn't break
the current workflow.

This flag should be used with the "Ns" flag to prevent a shutdown
provoked by a invalid file descriptor added after latter.

Signed-off-by: Arturo Volpe <avolpe@fintech.works>